api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg[binary]==3.1.13
python-dotenv==1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
